can now delete a customer

+++ customers/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render
+from django.shortcuts import render, redirect
 from django.http import HttpResponse
 from .forms import *
 
@@ -23,17 +23,19 @@ def create(request):
     customer_form = CustomerForm
     address_form = AddressForm
     car_form = CarForm
-    
+
     if request.method == 'POST':
         customer_form = CustomerForm(request.POST)
         if customer_form.is_valid():
             customer_form.save()
-    
+            return redirect('/')
+
     return render(request, 'customers/create.html', { "customer_form": customer_form, "address_form": address_form, "car_form": car_form })
 
 def delete(request, id):
     customer = Customer.objects.get(id=id)
-    # Need to fix here
-    customer.delete()
-  
+    if request.method == 'POST':
+      customer.delete()
+      return redirect('/')
+
     return render(request, 'customers/delete.html', { "customer": customer })
